[Chore/91] Notification Inactivated
